# Optimize summary queries