@cproensa 2019-03-11
I have tested mariadb, pgsql and oracle
Can't test mssql at the moment, as adodb+sqlsrv5.6+php7.3 is broken