@dregad 2019-04-14
Opened [#25693](https://www.mantisbt.org/bugs/view.php?id=25693) to track this.